LPS-69343 Move to DDLDDMBaseDataFactory and move the util method to InitContext class
