+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/InitContext.java
@@ -25,6 +25,7 @@ import com.liferay.util.SimpleCounter;
 import java.text.Format;
 
 import java.util.Collection;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -71,6 +72,18 @@ public class InitContext {
 
 	public SimpleCounter getCounter() {
 		return _counter;
+	}
+
+	public String getDateLong(Date date) {
+		return String.valueOf(date.getTime());
+	}
+
+	public String getDateString(Date date) {
+		if (date == null) {
+			return null;
+		}
+
+		return _simpleDateFormat.format(date);
 	}
 
 	public long getDefaultUserId() {
