+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -70,7 +70,7 @@ public class SampleSQLBuilder {
 
 			DataFactory dataFactory = new DataFactory(initContext);
 
-			new SampleSQLBuilder(properties, dataFactory);
+			new SampleSQLBuilder(properties, dataFactory, initContext);
 		}
 		catch (Exception e) {
 			e.printStackTrace();
@@ -87,7 +87,9 @@ public class SampleSQLBuilder {
 		}
 	}
 
-	public SampleSQLBuilder(Properties properties, DataFactory dataFactory)
+	public SampleSQLBuilder(
+			Properties properties, DataFactory dataFactory,
+			InitContext initContext)
 		throws Exception {
 
 		_dbType = DBType.valueOf(
@@ -102,6 +104,7 @@ public class SampleSQLBuilder {
 		_script = properties.getProperty("sample.sql.script");
 
 		_dataFactory = dataFactory;
+		_initContext = initContext;
 
 		// Generic
 
@@ -353,8 +356,7 @@ public class SampleSQLBuilder {
 	protected Map<String, Object> getContext() throws Exception {
 		Map<String, Object> context = _dataFactory.getDataFactories();
 
-		context.put("dataFactory", _dataFactory);
-		context.put("initContext", _dataFactory.getInitContext());
+		context.put("initContext", _initContext);
 
 		for (String csvFileName : _csvFileNames) {
 			Writer csvWriter = createFileWriter(
@@ -458,6 +460,7 @@ public class SampleSQLBuilder {
 	private final DataFactory _dataFactory;
 	private final DBType _dbType;
 	private volatile Throwable _freeMarkerThrowable;
+	private final InitContext _initContext;
 	private final int _optimizeBufferSize;
 	private final String _outputDir;
 	private final String _script;
