+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/asset.ftl
@@ -1,5 +1,5 @@
 <#list assetDataFactory.assetVocabularyModels as assetVocabularyModel>
-	insert into AssetVocabulary values ('${assetVocabularyModel.uuid}', ${assetVocabularyModel.vocabularyId}, ${assetVocabularyModel.groupId}, ${assetVocabularyModel.companyId}, ${assetVocabularyModel.userId}, '${assetVocabularyModel.userName}', '${dataFactory.getDateString(assetVocabularyModel.createDate)}', '${dataFactory.getDateString(assetVocabularyModel.modifiedDate)}', '${assetVocabularyModel.name}', '${assetVocabularyModel.title}', '${assetVocabularyModel.description}', '${assetVocabularyModel.settings}', '${dataFactory.getDateString(assetVocabularyModel.lastPublishDate)}');
+	insert into AssetVocabulary values ('${assetVocabularyModel.uuid}', ${assetVocabularyModel.vocabularyId}, ${assetVocabularyModel.groupId}, ${assetVocabularyModel.companyId}, ${assetVocabularyModel.userId}, '${assetVocabularyModel.userName}', '${initContext.getDateString(assetVocabularyModel.createDate)}', '${initContext.getDateString(assetVocabularyModel.modifiedDate)}', '${assetVocabularyModel.name}', '${assetVocabularyModel.title}', '${assetVocabularyModel.description}', '${assetVocabularyModel.settings}', '${initContext.getDateString(assetVocabularyModel.lastPublishDate)}');
 
 	<@insertResourcePermissions
 		_entry = assetVocabularyModel
@@ -7,7 +7,7 @@
 </#list>
 
 <#list assetDataFactory.assetCategoryModels as assetCategoryModel>
-	insert into AssetCategory values ('${assetCategoryModel.uuid}', ${assetCategoryModel.categoryId}, ${assetCategoryModel.groupId}, ${assetCategoryModel.companyId}, ${assetCategoryModel.userId}, '${assetCategoryModel.userName}', '${dataFactory.getDateString(assetCategoryModel.createDate)}', '${dataFactory.getDateString(assetCategoryModel.modifiedDate)}', ${assetCategoryModel.parentCategoryId}, ${assetCategoryModel.leftCategoryId}, ${assetCategoryModel.rightCategoryId}, '${assetCategoryModel.name}', '${assetCategoryModel.title}', '${assetCategoryModel.description}', ${assetCategoryModel.vocabularyId}, '${dataFactory.getDateString(assetCategoryModel.lastPublishDate)}');
+	insert into AssetCategory values ('${assetCategoryModel.uuid}', ${assetCategoryModel.categoryId}, ${assetCategoryModel.groupId}, ${assetCategoryModel.companyId}, ${assetCategoryModel.userId}, '${assetCategoryModel.userName}', '${initContext.getDateString(assetCategoryModel.createDate)}', '${initContext.getDateString(assetCategoryModel.modifiedDate)}', ${assetCategoryModel.parentCategoryId}, ${assetCategoryModel.leftCategoryId}, ${assetCategoryModel.rightCategoryId}, '${assetCategoryModel.name}', '${assetCategoryModel.title}', '${assetCategoryModel.description}', ${assetCategoryModel.vocabularyId}, '${initContext.getDateString(assetCategoryModel.lastPublishDate)}');
 
 	<@insertResourcePermissions
 		_entry = assetCategoryModel
@@ -15,7 +15,7 @@
 </#list>
 
 <#list assetDataFactory.assetTagModels as assetTagModel>
-	insert into AssetTag values ('${assetTagModel.uuid}', ${assetTagModel.tagId}, ${assetTagModel.groupId}, ${assetTagModel.companyId}, ${assetTagModel.userId}, '${assetTagModel.userName}', '${dataFactory.getDateString(assetTagModel.createDate)}', '${dataFactory.getDateString(assetTagModel.modifiedDate)}', '${assetTagModel.name}', ${assetTagModel.assetCount}, '${dataFactory.getDateString(assetTagModel.lastPublishDate)}');
+	insert into AssetTag values ('${assetTagModel.uuid}', ${assetTagModel.tagId}, ${assetTagModel.groupId}, ${assetTagModel.companyId}, ${assetTagModel.userId}, '${assetTagModel.userName}', '${initContext.getDateString(assetTagModel.createDate)}', '${initContext.getDateString(assetTagModel.modifiedDate)}', '${assetTagModel.name}', ${assetTagModel.assetCount}, '${initContext.getDateString(assetTagModel.lastPublishDate)}');
 
 	<@insertResourcePermissions
 		_entry = assetTagModel
