+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/blogs.ftl
@@ -1,7 +1,7 @@
 <#assign blogsEntryModels = blogDataFactory.newBlogsEntryModels(groupId) />
 
 <#list blogsEntryModels as blogsEntryModel>
-	insert into BlogsEntry values ('${blogsEntryModel.uuid}', ${blogsEntryModel.entryId}, ${blogsEntryModel.groupId}, ${blogsEntryModel.companyId}, ${blogsEntryModel.userId}, '${blogsEntryModel.userName}', '${dataFactory.getDateString(blogsEntryModel.createDate)}', '${dataFactory.getDateString(blogsEntryModel.modifiedDate)}', '${blogsEntryModel.title}', '${blogsEntryModel.subtitle}', '${blogsEntryModel.urlTitle}', '${blogsEntryModel.description}', '${blogsEntryModel.content}', '${dataFactory.getDateString(blogsEntryModel.displayDate)}', ${blogsEntryModel.allowPingbacks?string}, ${blogsEntryModel.allowTrackbacks?string}, '${blogsEntryModel.trackbacks}', '${blogsEntryModel.coverImageCaption}', ${blogsEntryModel.coverImageFileEntryId}, '${blogsEntryModel.coverImageURL}', ${blogsEntryModel.smallImage?string}, ${blogsEntryModel.smallImageFileEntryId}, ${blogsEntryModel.smallImageId}, '${blogsEntryModel.smallImageURL}', '${dataFactory.getDateString(blogsEntryModel.lastPublishDate)}', ${blogsEntryModel.status}, ${blogsEntryModel.statusByUserId}, '${blogsEntryModel.statusByUserName}', '${dataFactory.getDateString(blogsEntryModel.statusDate)}');
+	insert into BlogsEntry values ('${blogsEntryModel.uuid}', ${blogsEntryModel.entryId}, ${blogsEntryModel.groupId}, ${blogsEntryModel.companyId}, ${blogsEntryModel.userId}, '${blogsEntryModel.userName}', '${initContext.getDateString(blogsEntryModel.createDate)}', '${initContext.getDateString(blogsEntryModel.modifiedDate)}', '${blogsEntryModel.title}', '${blogsEntryModel.subtitle}', '${blogsEntryModel.urlTitle}', '${blogsEntryModel.description}', '${blogsEntryModel.content}', '${initContext.getDateString(blogsEntryModel.displayDate)}', ${blogsEntryModel.allowPingbacks?string}, ${blogsEntryModel.allowTrackbacks?string}, '${blogsEntryModel.trackbacks}', '${blogsEntryModel.coverImageCaption}', ${blogsEntryModel.coverImageFileEntryId}, '${blogsEntryModel.coverImageURL}', ${blogsEntryModel.smallImage?string}, ${blogsEntryModel.smallImageFileEntryId}, ${blogsEntryModel.smallImageId}, '${blogsEntryModel.smallImageURL}', '${initContext.getDateString(blogsEntryModel.lastPublishDate)}', ${blogsEntryModel.status}, ${blogsEntryModel.statusByUserId}, '${blogsEntryModel.statusByUserName}', '${initContext.getDateString(blogsEntryModel.statusDate)}');
 
 	<@insertResourcePermissions
 		_entry = blogsEntryModel
