+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/company.ftl
@@ -4,7 +4,7 @@ insert into Company values (${companyModel.mvccVersion}, ${companyModel.companyI
 
 <#assign accountModel = userDataFactory.accountModel />
 
-insert into Account_ values (${accountModel.mvccVersion}, ${accountModel.accountId}, ${accountModel.companyId}, ${accountModel.userId}, '${accountModel.userName}', '${dataFactory.getDateString(accountModel.createDate)}', '${dataFactory.getDateString(accountModel.modifiedDate)}', '${accountModel.parentAccountId}', '${accountModel.name}', '${accountModel.legalName}', '${accountModel.legalId}', '${accountModel.legalType}', '${accountModel.sicCode}', '${accountModel.tickerSymbol}', '${accountModel.industry}', '${accountModel.type}', '${accountModel.size}');
+insert into Account_ values (${accountModel.mvccVersion}, ${accountModel.accountId}, ${accountModel.companyId}, ${accountModel.userId}, '${accountModel.userName}', '${initContext.getDateString(accountModel.createDate)}', '${initContext.getDateString(accountModel.modifiedDate)}', '${accountModel.parentAccountId}', '${accountModel.name}', '${accountModel.legalName}', '${accountModel.legalId}', '${accountModel.legalType}', '${accountModel.sicCode}', '${accountModel.tickerSymbol}', '${accountModel.industry}', '${accountModel.type}', '${accountModel.size}');
 
 <#assign virtualHostModel = userDataFactory.virtualHostModel />
