+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/roles.ftl
@@ -1,5 +1,5 @@
 <#list userDataFactory.roleModels as roleModel>
-	insert into Role_ values (${roleModel.mvccVersion}, '${roleModel.uuid}', ${roleModel.roleId}, ${roleModel.companyId}, ${roleModel.userId}, '${roleModel.userName}', '${dataFactory.getDateString(roleModel.createDate)}', '${dataFactory.getDateString(roleModel.modifiedDate)}', ${roleModel.classNameId}, ${roleModel.classPK}, '${roleModel.name}', '${roleModel.title}', '${roleModel.description}', ${roleModel.type}, '${roleModel.subtype}');
+	insert into Role_ values (${roleModel.mvccVersion}, '${roleModel.uuid}', ${roleModel.roleId}, ${roleModel.companyId}, ${roleModel.userId}, '${roleModel.userName}', '${initContext.getDateString(roleModel.createDate)}', '${initContext.getDateString(roleModel.modifiedDate)}', ${roleModel.classNameId}, ${roleModel.classPK}, '${roleModel.name}', '${roleModel.title}', '${roleModel.description}', ${roleModel.type}, '${roleModel.subtype}');
 
 	<@insertResourcePermissions
 		_entry = roleModel
